// frontend/src/app/interfaces/SalesMan.ts
export interface SalesMan {
    firstName: String;
    middleName: String;
    lastName: String;
    code: String;
    employeeId: String;
    fullName: String;
    status: String;
    dob: String;
    driversLicenseNumber: String;
    licenseExpiryDate: String;
    maritalStatus: String;
    gender: String;
    otherId: String;
    nationality: String;
    unit: String;
    jobTitle: String;
    supervisor: SalesMan;
}